integrity: Switch from rbtree to LSM-managed blob for integrity_iint_cache
Before the security field of kernel objects could be shared among LSMs with
the LSM stacking feature, IMA and EVM had to rely on an alternative storage
of inode metadata. The association between inode metadata and inode is
maintained through an rbtree.

With the reservation mechanism offered by the LSM infrastructure, the
rbtree is no longer necessary, as each LSM could reserve a space in the
security blob for each inode.

With the 'integrity' LSM removed, reserve space from IMA (or EVM if IMA is
disabled) in the inode security blob for a pointer to the
integrity_iint_cache structure directly, rather than embedding the whole
structure in the inode security blob, to minimize the changes and to avoid
waste of memory.

Introduce two primitives for getting and setting the pointer of
integrity_iint_cache in the security blob, respectively
integrity_inode_get_iint() and integrity_inode_set_iint(). This would make
the code more understandable, as they directly replace rbtree operations.

Locking is not needed, as access to inode metadata is not shared, it is per
inode.

Finally, with integrity_blob_sizes defined, switch to the LSM reservation
mechanism for the EVM xattr, and consequently decrement by one the number
of xattrs to allocate in security_inode_init_security().

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>