digest_cache: Reset digest cache on file/directory change
Register seven new LSM hooks, file_open, path_truncate, file_release,
inode_unlink, inode_rename, inode_post_setxattr and inode_post_removexattr,
to monitor digest lists/directory modifications.

If an action affects a digest list or the parent directory, the new LSM
hook implementations call digest_cache_reset_owner() to set the RESET bit
on the digest cache pointed to by dig_owner in the inode security blob.
This will cause next calls to digest_cache_get() and digest_cache_create()
to respectively put and clear dig_user and dig_owner, and request a new
digest cache.

If an action affects a file using a digest cache,  the new LSM hook
implementations call digest_cache_clear_user() to release the digest cache
reference pointed to by dig_user in the inode security blob and to clear
dig_user.

That does not affect other users of the old digest cache, since that one
remains valid as long as the reference count is greater than zero. However,
they can explicitly call the new function digest_cache_changed(), to check
if, or dig_user was released and cleared, or the RESET bit was set (the
digest cache will be released at the next digest_cache_get() call).

Recreating a file digest cache means reading the digest list again and
extracting the digests. Recreating a directory digest cache, instead, does
not mean recreating the digest cache for directory entries, since those
digest caches are likely already stored in the inode security blob. It
would happen however for new files.

File digest cache dig_owner reset is done on file_open, when a digest list
is opened for write, path_truncate, when a digest list is truncated (there
is no inode_truncate, file_truncate does not catch operations through the
truncate() system call), inode_unlink, when a digest list is removed, and
inode_rename when a digest list is renamed.

Directory digest cache dig_owner reset is done on file_release, when a
digest list is written in the digest list directory, on inode_unlink, when
a digest list is deleted from that directory, and finally on inode_rename,
when a digest list is moved to/from that directory.

File digest cache dig_user clear is done on inode_post_setxattr and
inode_post_removexattr, when the security.digest_list xattr is respectively
set or removed from a file using that digest cache.

With the exception of file_release, which will always be executed (cannot
be denied), and inode_post_setxattr and inode_post_removexattr, which are
executed after the actual operation, the other LSM hooks are not optimal,
since the digest_cache LSM does not know whether or not the operation will
be allowed also by other LSMs. If the operation is denied, the digest_cache
LSM would do an unnecessary reset.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>