ima: Ensure lock is held when setting iint pointer in inode security blob
IMA stores a pointer of the ima_iint_cache structure, containing integrity
metadata, in the inode security blob. However, check and assignment of this
pointer is not atomic, and it might happen that two tasks both see the iint
pointer to NULL and try to set it.

Overcome this issue by guarding the check and iint assignment in
ima_iint_get() with the iint lock, and by adding a lockdep assertion, to
ensure that the lock is actually held.

Consequently, guard the remaining ima_iint_get() calls, in
ima_post_create_tmpfile() and ima_post_path_mknod().

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>