ima: Discard files opened with O_PATH
According to the man, files opened with O_PATH are not really opened. The
obtained file descriptor is used to indicate a location in the filesystem
tree and to perform operations that act purely at the file descriptor
level.

Thus, ignore open() syscalls with O_PATH, since IMA cares about file data.

Cc: stable@vger.kernel.org # v2.6.39.x
Fixes: 1abf0c718f15a ("New kind of open files - "location only".")
Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>