ima: Attach CREDS_CHECK IMA hook to bprm_creds_from_file LSM hook
Since commit 56305aa9b6fa ("exec: Compute file based creds only once"), the
credentials to be applied to the process after execution are not calculated
anymore for each step of finding intermediate interpreters, but only after
the final binary to be executed without interpreter has been found.

As a consequence, the bprm_check_security LSM hook is not anymore suitable
for invoking the CREDS_CHECK hook, since the latter expects to see the
ready-to-be-applied credentials instead of the partially set ones (LSMs set
their final credentials in the bprm_creds_for_exec hook, DAC set them
later).

By invoking CREDS_CHECK in the bprm_check_security hook, an IMA policy for
example with euid=0 will not be matched for set[ug]id programs, since the
credentials seen at this point are the ones of the regular users launching
that program.

Fix this issue by moving the CREDS_CHECK hook from the bprm_check_security
LSM hook to bprm_creds_from_file, from which the final credentials can be
obtained. Compared to before the commit mentioned above, CREDS_CHECK will
not be invoked anymore for intermediate interpreters.

Link: https://github.com/linux-integrity/linux/issues/3
Fixes: 56305aa9b6fa ("exec: Compute file based creds only once")
Cc: Serge E. Hallyn <serge@hallyn.com>
Cc: Matthew Garrett <mjg59@srcf.ucam.org>
Cc: Eric W. Biederman <ebiederm@xmission.com>
Cc: Jann Horn <jannh@google.com>
Cc: Christian Brauner <brauner@kernel.org>
Cc: Kees Cook <keescook@chromium.org>
Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>