ima: Attach CREDS_CHECK IMA hook to bprm_creds_from_file LSM hook
Since commit 56305aa9b6fa ("exec: Compute file based creds only once"), the
credentials to be applied to the process after execution are not calculated
anymore for each step of finding intermediate interpreters (including the
final binary), but only after the final binary to be executed without
interpreter has been found.

In particular, that means that the bprm_check_security LSM hook will not
see the updated cred->e[ug]id for the intermediate and for the final binary
to be executed, since the function doing this task has been moved from
prepare_binprm(), which calls the bprm_check_security hook, to
bprm_creds_from_file().

This breaks the IMA expectation for the CREDS_CHECK hook, introduced with
commit d906c10d8a31 ("IMA: Support using new creds in appraisal policy"),
which expects to evaluate "the credentials that will be committed when the
new process is started", which is clearly not the case for
bprm_check_security the CREDS_CHECK IMA hook is attached to.

This issue does not affect systems with the default IMA policy loaded,
since the default measurement policy measures all executions regardless of
the new credentials, and the default appraisal policy is written in terms
of the file owner, rather than IMA hooks.

However, it does affect systems with a policy where the BPRM_CHECK hook
does not include what CREDS_CHECK evaluates. As an extreme example, taking
a standalone rule like:

measure func=CREDS_CHECK euid=0

this will not measure for example sudo (because CREDS_CHECK still sees the
bprm->cred->euid set to the regular user UID), but only the subsequent
commands after the euid was applied to the children.

Since currently there is no mechanism in the IMA policy code to enforce
BPRM_CHECK to be inclusive of CRED_CHECK hook, make set[ug]id programs
measured/appraised again by splitting ima_bprm_check() in two separate
hook implementations (CREDS_CHECK now being implemented by
ima_creds_check()), and by attaching CREDS_CHECK to the
bprm_creds_from_file LSM hook.

The limitation of this approach is that CREDS_CHECK will not be invoked
anymore for the intermediate interpreters but only for the final binary.
This limitation can be removed only by moving bprm_fill_uid() back to
where it was in prepare_bprm().

Link: https://github.com/linux-integrity/linux/issues/3
Fixes: 56305aa9b6fa ("exec: Compute file based creds only once")
Cc: Serge E. Hallyn <serge@hallyn.com>
Cc: Matthew Garrett <mjg59@srcf.ucam.org>
Cc: Eric W. Biederman <ebiederm@xmission.com>
Cc: Jann Horn <jannh@google.com>
Cc: Christian Brauner <brauner@kernel.org>
Cc: Kees Cook <keescook@chromium.org>
Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>