Add a measurement entry only if everything else succeeded
Commit 7b7e5916aa2f4 ("ima: fix invalid memory reference") solved the
problem of invalid reference by not passing the TPM error to the callers of
ima_template_add_entry().

That could have worked at that time, since errors were only logged, and the
caller only used that error to free the measurement entry.

With the upcoming change of passing the error down to the LSM security
framework, the TPM error needs to be passed to the caller too. However, to
avoid the problem fixed by the above commit, add a measurement entry as the
last operation before returning, so that there is no risk that the caller
frees an already added one.

Filter out the -EEXIST error from the TPM (change it to -EIO), since that
error currently means that the measurement operation succeeded, but there
is a digest collision between the new entry and the already existing
entries (and duplicate entries are not allowed in the kernel
configuration).

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>