ima: Deny file op if a measurement cannot be made
Traditionally, taking integrity measurements should be done without
interfering with the systems. Taking measurement before an operation is
performed, and the binding of a TPM PCR extend operation with the previous
PCR value ensure that a system compromise can be detected after the fact
without denying the operation, as secure boot does.

However, this scheme works as long as a measurement is recorded with the
TPM. If for any reason, an error occurs during the measurement, so that
the measurement is not recorded with the TPM, IMA cannot let the operation
be performed, because otherwise system integrity would be evaluated from
incomplete information, potentially with the measurement revealing an
attack missing.

Thus, propagate the error down from ima_template_add_entry() to
process_measurement() and process_buffer_measurement(). If an error occurs,
return an error to the LSM security frameworks, so that the operation is
denied.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>