ima: Attempt to invalidate the measurement list before denying operation
Denying an operation if only a measurement is requested is not the optimal
solution, since measurement is not supposed to do that.

Introduce ima_invalidate_measurement_list() as a last attempt to avoid
returning an error to user space. Calling this function causes the
measurement list to be invalidated, because it performs a PCR extend
without adding a new measurement entry (which would likely incur in the
same error as the first attempt).

However, if it succeeds, it achieve the purpose of letting the remote
verifiers know that the measurement list is incomplete and should be
discarded. Thus, if it succeeds, don't pass the error to user space.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>